Fixed gitfs cachedir_basename to avoid hash collisions
