+++ changelog/cve-2023-20898.security.md
@@ -1,0 +1,1 @@
+Fixed gtfs cachedir_basename to avoid hash collisions. Added MP Lock to gtfs. These changes should stop race conditions.
